base area
git-svn-id: https://webgate.ec.europa.eu/CITnet/svn/UNIONVMS/trunk/Modules/Spatial@11380 ffa9caa6-9a2d-11e4-8c71-fd203efb3018

+++ APP/service/src/test/java/eu/europa/ec/fisheries/uvms/spatial/dao/UserAreaDaoTest.java
@@ -46,7 +46,7 @@ public class UserAreaDaoTest extends BaseSpatialDaoTest {
     @SneakyThrows
     public void testFindByUserNameAndSCopeName(){
         List<UserAreasEntity> entityList = dao.findByUserNameAndScopeName("userDaoTest", "EC");
-        assertEquals(1L, entityList.get(0).getGid());
+        assertEquals(1L, entityList.get(0).getGid().longValue());
     }
